pyproject.toml
[project]
name = "mcp-manager"
version = "0.1.0"
description = "Centralized MCP server management and routing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "anyio>=4.0.0",
    "tabulate>=0.9.0",
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
]

[project.scripts]
mcp-manager = "mcp_manager.cli:main"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
